telco-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "telco-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "bnachlieli", email = "bnachlieli@nvidia.com" }]
requires-python = ">=3.12.9"
dependencies = ["fastapi", "uvicorn", "pydantic-settings"]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "pytest-cov", "flake8", "mypy"]

[project.scripts]
telco-service = "telco_service.__main__:main"
